backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.6.1
PyMuPDF==1.23.26
chromadb==0.5.23
sentence-transformers==2.3.1
openai==1.12.0
httpx==0.27.0
langchain==0.1.6
langchain-community==0.0.19
langchain-openai==0.0.5
reportlab==4.0.9
Pillow==10.2.0
numpy==1.26.4
pandas==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
email-validator==2.1.0
